Exclude tranching liquidity from core TVL
